# docs/swagger.yaml
basePath: /api/v1
definitions:
  handlers.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  handlers.PaginatedResponse:
    properties:
      data: {}
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
  models.CreateReservationRequest:
    properties:
      customer_email:
        type: string
      customer_name:
        maxLength: 100
        minLength: 2
        type: string
      customer_phone:
        type: string
      date:
        type: string
      guests:
        maximum: 20
        minimum: 1
        type: integer
      special_requests:
        type: string
      status:
        allOf:
        - $ref: '#/definitions/models.ReservationStatus'
        enum:
        - pending
        - confirmed
        - cancelled
      time:
        type: string
      user_id:
        type: string
    required:
    - customer_email
    - customer_name
    - customer_phone
    - date
    - guests
    - time
    type: object
...
  models.ReservationResponse:
    properties:
      created_at:
        type: string
      customer_email:
        type: string
      customer_name:
        type: string
      customer_phone:
        type: string
      date:
        type: string
      guests:
        type: integer
      id:
        type: string
      special_requests:
        type: string
      status:
        $ref: '#/definitions/models.ReservationStatus'
      time:
        type: string
      updated_at:
        type: string
      user:
        $ref: '#/definitions/models.UserResponse'
      user_id:
        type: string
    type: object
  models.ReservationStatus:
    enum:
    - pending
    - confirmed
    - cancelled
    type: string
    x-enum-varnames:
    - ReservationStatusPending
    - ReservationStatusConfirmed
    - ReservationStatusCancelled
...
  models.UpdateReservationRequest:
    properties:
      customer_email:
        type: string
      customer_name:
        maxLength: 100
        minLength: 2
        type: string
      customer_phone:
        type: string
      date:
        type: string
      guests:
        maximum: 20
        minimum: 1
        type: integer
      special_requests:
        type: string
      status:
        allOf:
        - $ref: '#/definitions/models.ReservationStatus'
        enum:
        - pending
        - confirmed
        - cancelled
      time:
        type: string
    type: object
...
  models.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      last_login:
        type: string
      name:
        type: string
      phone:
        type: string
      role:
        $ref: '#/definitions/models.UserRole'
      status:
        $ref: '#/definitions/models.UserStatus'
      updated_at:
        type: string
      username:
        type: string
    type: object
  models.UserRole:
    enum:
    - admin
    - manager
    - staff
    type: string
    x-enum-varnames:
    - RoleAdmin
    - RoleManager
    - RoleStaff
  models.UserStatus:
    enum:
    - active
    - inactive
    type: string
    x-enum-varnames:
    - StatusActive
    - StatusInactive
host: localhost:8080
info:
  contact:
    email: support@vibandavillage.com
    name: API Support
    url: http://www.swagger.io/support
  description: A comprehensive backend API for Vibanda Village restaurant management
    system
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Vibanda Village Admin API
  version: "1.0"
paths:
...
  /reservations:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all reservations with pagination
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Search term
        in: query
        name: search
        type: string
      - description: Filter by status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.PaginatedResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all reservations
      tags:
      - reservations
    post:
      consumes:
      - application/json
      description: Create a new reservation
      parameters:
      - description: Reservation data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateReservationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.ReservationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new reservation
      tags:
      - reservations
  /reservations/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a reservation
      parameters:
      - description: Reservation ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete reservation
      tags:
      - reservations
    get:
      consumes:
      - application/json
      description: Retrieve a specific reservation by ID
      parameters:
      - description: Reservation ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ReservationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get reservation by ID
      tags:
      - reservations
    put:
      consumes:
      - application/json
      description: Update an existing reservation
      parameters:
      - description: Reservation ID
        in: path
        name: id
        required: true
        type: string
      - description: Reservation update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateReservationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ReservationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update reservation
      tags:
      - reservations
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"